# openapi/butly.openapi.json
{
  "components": {
    "schemas": {
      "ApiError": {
        "description": "`/api/v1` の共通 error envelope。\n\nHTTP response では常に 4 field すべてを含める（default 省略はしない）。",
        "properties": {
          "code": {
            "description": "frontend 分岐用の安定した error code（snake_case）",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "description": "error code ごとの補足情報（validation error の一覧など）",
            "title": "Details",
            "type": "object"
          },
          "message": {
            "description": "人間可読なエラー説明。secret や内部 path を含めない",
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "server log と突合するための request ID（X-Request-ID と同値）",
            "title": "Request Id"
          }
        },
        "required": [
          "code",
          "details",
          "message",
          "request_id"
        ],
        "title": "ApiError",
        "type": "object"
      },
...
      "ChatDebugSummary": {
        "properties": {
          "gatekeeper": {
            "$ref": "#/components/schemas/GatekeeperDebugSummary"
          },
          "rag": {
            "$ref": "#/components/schemas/RagDebugSummary"
          }
        },
        "required": [
          "gatekeeper"
        ],
        "title": "ChatDebugSummary",
        "type": "object"
      },
...
      "ChatRequest": {
        "additionalProperties": false,
        "description": "`POST /api/v1/chat` / `/api/v1/chat/stream` 共通 request。\n\n- `use_google_search` と `use_web_search` は相互排他（backend で validation）。\n- `model_name` / `connection` は developer override。通常 UI は\n  instance config に従い未指定にする。\n- unknown field は拒否し typo を早期検出する（legacy route のみ互換を緩める）。",
        "properties": {
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/AttachmentInput"
            },
            "maxItems": 3,
            "title": "Attachments",
            "type": "array"
          },
          "client_request_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "SSE の double submit/replay 識別用に client が採番する ID。non-stream fallback は idempotency 対象外",
            "title": "Client Request Id"
          },
          "connection": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Connection"
          },
          "include_debug": {
            "default": false,
            "description": "developer mode で安全な Gatekeeper/RAG 要約を含める",
            "title": "Include Debug",
            "type": "boolean"
          },
          "instance_name": {
            "minLength": 1,
            "title": "Instance Name",
            "type": "string"
          },
          "model_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name"
          },
          "text": {
            "default": "",
            "title": "Text",
            "type": "string"
          },
          "use_google_search": {
            "default": false,
            "title": "Use Google Search",
            "type": "boolean"
          },
          "use_rag": {
            "default": true,
            "title": "Use Rag",
            "type": "boolean"
          },
          "use_web_search": {
            "default": false,
            "title": "Use Web Search",
            "type": "boolean"
          }
        },
        "required": [
          "instance_name"
        ],
        "title": "ChatRequest",
        "type": "object"
      },
      "ChatRequestStatus": {
        "description": "process-local generation request の状態。",
        "properties": {
          "attempt": {
            "default": 1,
            "minimum": 1.0,
            "title": "Attempt",
            "type": "integer"
          },
          "cancellable": {
            "default": false,
            "title": "Cancellable",
            "type": "boolean"
          },
          "client_request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Request Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              },
              {
                "type": "null"
              }
            ]
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "state": {
            "enum": [
              "running",
              "finalizing",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "state",
          "created_at"
        ],
        "title": "ChatRequestStatus",
        "type": "object"
      },
      "ChatResult": {
        "description": "`POST /api/v1/chat`（non-stream fallback / test 用）の応答。",
        "properties": {
          "debug": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChatDebugSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "keywords": {
            "items": {
              "type": "string"
            },
            "title": "Keywords",
            "type": "array"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/CitationSource"
            },
            "title": "Sources",
            "type": "array"
          },
          "status": {
            "const": "completed",
            "default": "completed",
            "title": "Status",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "text"
        ],
        "title": "ChatResult",
        "type": "object"
      },
      "ChatStreamEvent": {
        "discriminator": {
          "mapping": {
            "chunk": "#/components/schemas/ChatChunkEvent",
            "done": "#/components/schemas/ChatDoneEvent",
            "error": "#/components/schemas/ChatErrorEvent",
            "metadata": "#/components/schemas/ChatMetadataEvent"
          },
          "propertyName": "event"
        },
        "oneOf": [
          {
            "$ref": "#/components/schemas/ChatMetadataEvent"
          },
          {
            "$ref": "#/components/schemas/ChatChunkEvent"
          },
          {
            "$ref": "#/components/schemas/ChatDoneEvent"
          },
          {
            "$ref": "#/components/schemas/ChatErrorEvent"
          }
        ]
      },
      "CitationSource": {
        "description": "Web 検索 / grounding の引用元。provider raw object は正規化して返す。",
        "properties": {
          "title": {
            "default": "",
            "title": "Title",
            "type": "string"
          },
          "url": {
            "default": "",
            "title": "Url",
            "type": "string"
          }
        },
        "title": "CitationSource",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "DesktopToken": {
        "description": "Per-launch desktop token (BUTLY_DESKTOP_TOKEN). Not enforced when the backend runs without a token (dev mode).",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Butly backend API. Frontend-facing contract is under /api/v1.",
    "title": "Butly API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/chat": {
      "post": {
        "description": "チャットメッセージを送信し、完成した応答を返す。primary transport は `POST /api/v1/chat/stream`（SSE）で、この endpoint は non-stream fallback / test 用。legacy `POST /chat` の versioned 置換。",
        "operationId": "send_chat",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Developer debug is disabled"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Instance not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Idempotency conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Validation error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Request registry capacity exceeded"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Backend is not ready"
          }
        },
        "security": [
          {
            "DesktopToken": []
          }
        ],
        "summary": "Send a chat message (non-streaming)",
        "tags": [
          "chat"
        ]
      }
    },
    "/api/v1/chat/requests/{request_id}": {
      "get": {
        "operationId": "get_chat_request_status",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatRequestStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Request not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "DesktopToken": []
          }
        ],
        "summary": "Get generation request status",
        "tags": [
          "chat"
        ]
      }
    },
    "/api/v1/chat/requests/{request_id}/cancel": {
      "post": {
        "description": "実行中の server task をキャンセルし、対応する SDK では provider stream のclose も試みる。同期 SDK が close を提供しない場合は provider 側の処理が継続し得るが、部分応答は保存しない。保存開始後（finalizing）の cancel は拒否され、1回だけ保存して completed になる。terminal request への再実行はidempotent。",
        "operationId": "cancel_chat_request",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatRequestStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Request not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "DesktopToken": []
          }
        ],
        "summary": "Cancel generation",
        "tags": [
          "chat"
        ]
      }
    },
    "/api/v1/chat/stream": {
      "post": {
        "description": "チャットメッセージを送信し、`text/event-stream` で metadata → chunk* → done（失敗時は error 終端）を返す。event payload は components の `ChatStreamEvent`（oneOf + discriminator）。legacy `POST /chat/stream` の versioned 置換。",
        "operationId": "stream_chat",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ChatStreamEvent"
                }
              }
            },
            "description": "SSE stream of ChatStreamEvent frames"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Developer debug is disabled"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Instance not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Idempotency conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Validation error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Request registry capacity exceeded"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Backend is not ready"
          }
        },
        "security": [
          {
            "DesktopToken": []
          }
        ],
        "summary": "Send a chat message (SSE streaming)",
        "tags": [
          "chat"
        ]
      }
    },
...
  }
}